+++ Twinder/DeepLinks.swift
@@ -7,12 +7,17 @@
 //
 
 import Foundation
+import UIKit
 
 func twtrScreenName(screen_name: String){
-    let deepLinkUrl: NSString = "twitter://user?screen_name=" + screen_name
-    let url = NSURL(string: (deepLinkUrl))
-    if UIApplication.sharedApplication().canOpenURL(url!) {
-        // url is good to go
-        UIApplication.sharedApplication().openURL(url!)
+    let components = NSURLComponents()
+    components.scheme = "twitter"
+    components.host = "user"
+    components.queryItems = [NSURLQueryItem(name: "screen_name", value: screen_name)]
+
+    if let url = components.URL {
+        if UIApplication.sharedApplication().canOpenURL(url) {
+            UIApplication.sharedApplication().openURL(url)
+        }
     }
 }